test: NewsController 테스트코드
