+++ src/test/java/org/myteam/server/ControllerTestSupport.java
@@ -1,10 +1,7 @@
 package org.myteam.server;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.persistence.EntityManager;
-import org.myteam.server.article.controller.ArticleController;
-import org.myteam.server.article.service.ArticleReadService;
-import org.myteam.server.article.service.ArticleService;
+import org.myteam.server.news.controller.NewsController;
+import org.myteam.server.news.service.NewsReadService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -12,21 +9,23 @@ import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 @ActiveProfiles("test")
 @WebMvcTest(controllers = {
-    ArticleController.class,
+	NewsController.class,
 })
 @MockBean(JpaMetamodelMappingContext.class)
 public abstract class ControllerTestSupport {
 
-    @Autowired
-    protected MockMvc mockMvc;
+	@Autowired
+	protected MockMvc mockMvc;
 
-    @MockBean
-    protected ArticleService articleService;
+	@Autowired
+	protected ObjectMapper objectMapper;
 
-    @MockBean
-    protected ArticleReadService articleReadService;
+	@MockBean
+	protected NewsReadService newsReadService;
 
 }
