+++ src/test/java/org/myteam/server/news/controller/NewsControllerTest.java
@@ -1,0 +1,95 @@
+package org.myteam.server.news.controller;
+
+import static org.myteam.server.global.web.response.ResponseStatus.*;
+import static org.springframework.http.MediaType.*;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.myteam.server.ControllerTestSupport;
+import org.myteam.server.news.dto.controller.request.NewsRequest;
+import org.myteam.server.news.entity.NewsCategory;
+import org.myteam.server.news.repository.OrderType;
+import org.springframework.security.test.context.support.WithMockUser;
+
+class NewsControllerTest extends ControllerTestSupport {
+
+	@DisplayName("뉴스목록을 조회한다.")
+	@Test
+	@WithMockUser
+	void findAllTest() throws Exception {
+		// given
+		NewsRequest newsRequest = NewsRequest.builder()
+			.page(1)
+			.size(10)
+			.orderType(OrderType.LIKE)
+			.category(NewsCategory.FOOTBALL)
+			.build();
+
+		// when // then
+		mockMvc.perform(
+				get("/api/news")
+					.content(objectMapper.writeValueAsString(newsRequest))
+					.contentType(APPLICATION_JSON)
+					.accept(APPLICATION_JSON)
+					.with(csrf())
+			)
+			.andDo(print())
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("$.status").value(SUCCESS.name()))
+			.andExpect(jsonPath("$.msg").value("뉴스 목록 조회 성공"));
+	}
+
+	@DisplayName("뉴스목록을 조회시 정렬 타입은 필수이다.")
+	@Test
+	@WithMockUser
+	void findAllWithoutOrderTypeTest() throws Exception {
+		// given
+		NewsRequest newsRequest = NewsRequest.builder()
+			.page(1)
+			.size(10)
+			.category(NewsCategory.FOOTBALL)
+			.build();
+
+		// when // then
+		mockMvc.perform(
+				get("/api/news")
+					.content(objectMapper.writeValueAsString(newsRequest))
+					.contentType(APPLICATION_JSON)
+					.accept(APPLICATION_JSON)
+					.with(csrf())
+			)
+			.andDo(print())
+			.andExpect(status().isBadRequest())
+			.andExpect(jsonPath("$.status").value("BAD_REQUEST"))
+			.andExpect(jsonPath("$.message").value("뉴스 정렬 타입은 필수입니다."));
+	}
+
+	@DisplayName("뉴스목록을 조회한다.")
+	@Test
+	@WithMockUser
+	void findAllWithoutCategoryTest() throws Exception {
+		// given
+		NewsRequest newsRequest = NewsRequest.builder()
+			.page(1)
+			.size(10)
+			.orderType(OrderType.LIKE)
+			.build();
+
+		// when // then
+		mockMvc.perform(
+				get("/api/news")
+					.content(objectMapper.writeValueAsString(newsRequest))
+					.contentType(APPLICATION_JSON)
+					.accept(APPLICATION_JSON)
+					.with(csrf())
+			)
+			.andDo(print())
+			.andExpect(status().isBadRequest())
+			.andExpect(jsonPath("$.status").value("BAD_REQUEST"))
+			.andExpect(jsonPath("$.message").value("뉴스 카테고리는 필수입니다."));
+	}
+}
